2nd commit

+++ src/main/java/com/open/suite/controller/CommentController.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 
-import javax.servlet.http.HttpSession;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -16,10 +16,18 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.open.suite.domain.Comment;
 import com.open.suite.domain.ThreadRegistration;
+import com.open.suite.domain.User;
 import com.open.suite.service.CommentHql;
 import com.open.suite.service.CommentService;
 import com.open.suite.service.ThreadRegistrationService;
+import com.open.suite.service.UserService;
 import com.open.suite.util.DateUtil;
+/**
+ * @author    Minhajul Sarkar<polash.pbt@gmail.com>
+ * @version   1.0.00
+ * @since     1.0.00
+ * 
+ */
 
 @Controller
 @RequestMapping("/online/forum/comment")
@@ -35,18 +43,16 @@ public class CommentController {
 	@Autowired 
 	private CommentService commentService;
 	
+	@Autowired 
+	private UserService userService;
+	
 	String searchValue="";
 	@RequestMapping()
-	public String index(HttpSession session, ModelMap modelMap, String threadCode){
-		/*@SuppressWarnings("unused")
-		String userCode = (String) session.getAttribute("USER_CODE");
-		userCode=null;*/
+	public String index(ModelMap modelMap, String threadCode){
 		ThreadRegistration threadObj=service.findThreadByCode(threadCode);
 		modelMap.put("title", "Title : "+threadObj.getThreadTitle());
 		modelMap.put("description", "Description : "+threadObj.getThreadDescription());
 		modelMap.put("threadCode", threadCode);
-		
-		
 		searchValue=threadCode;
 		return "comment/index";
 	}
@@ -54,11 +60,13 @@ public class CommentController {
 	
 	@ResponseBody
 	@RequestMapping(method=RequestMethod.POST)
-	public HashMap<String, Object> save(HttpSession session, @RequestBody Comment jsonObj) {
+	public HashMap<String, Object> save(@RequestBody Comment jsonObj) {
 
 		HashMap<String, Object> map=new HashMap<String, Object>();
 		try {
-			
+			Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+			User user = userService.findUserByEmail(auth.getName());
+			if(user!=null){
 			Comment comment=new Comment();
 			comment.setCommentText(jsonObj.getCommentText());
 			comment.setCommentCode(DateUtil.format(new Date(), DateUtil.CODE_DATE_FORMAT));
@@ -68,6 +76,10 @@ public class CommentController {
 			commentService.save(comment);
 			map.put("message", "success");
 			map.put("tag", "info");
+		}else{
+			map.put("message", "not permission");
+			map.put("tag", "warn");
+		}
 		} catch (Exception e) {
 			/*map.put("message", messageByLocalService.getErrorMessage());
 			map.put("tag", "danger");*/
@@ -77,17 +89,19 @@ public class CommentController {
 	
 	
 	@RequestMapping(value="/delete", method=RequestMethod.GET)
-	public String delete(HttpSession session, String commentCode) {
+	public String delete(String commentCode) {
 
 		try {
+			Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+			User user = userService.findUserByEmail(auth.getName());
+			if(user!=null){
 			Comment comment=new Comment();
 			comment.setCommentCode(commentCode);
 			commentService.delete(comment);
+			}
 		} catch (Exception e) {
-			/*map.put("message", messageByLocalService.getErrorMessage());
-			map.put("tag", "danger");*/
+			e.printStackTrace();
 		}
-//		return "redirect:/opensuite/login";
 		return "redirect:/online/forum/comment?threadCode="+searchValue;
 	}
 	
@@ -95,25 +109,30 @@ public class CommentController {
 	
 	@ResponseBody
 	@RequestMapping(value ="/all", method=RequestMethod.GET)
-	public ArrayList<Comment> fetchAll(HttpSession session) {
+	public ArrayList<Comment> fetchAll() {
 		return serviceHql.fetchAllByCriteria(searchValue);
 	}
 	
 	@ResponseBody
 	@RequestMapping(value="/update", method=RequestMethod.POST)
-	public HashMap<String, Object> update(HttpSession session, @RequestBody Comment jsonObj) {
+	public HashMap<String, Object> update(@RequestBody Comment jsonObj) {
 		HashMap<String, Object> map=new HashMap<String, Object>();
 		try {
-			Comment commentObj=commentService.findByCommentCode(jsonObj.getCommentCode());
-			commentObj.setCommentText(jsonObj.getCommentText());
-			commentObj.setUpdateUser(commentObj.getEntryUser());
-			commentObj.setUpdateDate(new Date());
-			
-			jsonObj.setCommentCode(commentObj.getCommentCode());
-			
-			commentService.save(commentObj);
-			map.put("message", "Success");
-			map.put("tag", "info");
+			Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+			User user = userService.findUserByEmail(auth.getName());
+			if (user != null) {
+				Comment commentObj = commentService.findByCommentCode(jsonObj.getCommentCode());
+				commentObj.setCommentText(jsonObj.getCommentText());
+				commentObj.setUpdateUser(commentObj.getEntryUser());
+				commentObj.setUpdateDate(new Date());
+				jsonObj.setCommentCode(commentObj.getCommentCode());
+				commentService.save(commentObj);
+				map.put("message", "Success");
+				map.put("tag", "info");
+			}else{
+				map.put("message", "not permission");
+				map.put("tag", "warn");
+			}
 		} catch (Exception e) {
 			map.put("message", "Erro");
 			map.put("tag", "danger");
